### 2020-07-04 #2438 @whedon
[PRE REVIEW]: sonipy: A Perceptually Uniform Sonification Package
**Submitting author:** @lockepatton (<a href="http://orcid.org/0000-0002-7640-236X">Locke Patton</a>)
**Repository:** <a href="https://github.com/lockepatton/sonipy" target ="_blank">https://github.com/lockepatton/sonipy</a>
**Version:** v1.0
