### 2020-07-06 #2438 @harpolea
Hi @lockepatton, thanks for your submission!

Do you have any suggestions for potential reviewers? If so, then please mention them here in this thread (without tagging them with an @). In addition, this [list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
