### 2020-07-06 #2438 @lockepatton
Hi @harpolea ! Thanks for volunteering to be the editor. Looking through the list, there are three people that stood out. On the audio / signal processing side, there's **expectopatronum** and **faroit**. On the astronomy side, there's **benjaminrose** who works with supernova. I'd lean towards an audio/signal processing reviewer as the astronomy side is purely motivation, but any/all would work! My background is in astronomy, so any audio folks would have a fresh perspective. Cheers.
