### 2020-07-06 #2438 @harpolea
Great, thanks!

:wave: @expectopatronum & @faroit, would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
