### 2020-07-07 #2446 @whedon
[REVIEW]: soni-py: A Pitch-Based Sonification Package
**Submitting author:** @lockepatton (<a href="http://orcid.org/0000-0002-7640-236X">Locke Patton</a>)
**Repository:** https://github.com/lockepatton/sonipy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
