### 2020-07-08 8c827c5
Added binder link for interactive tutorial notebook
