### 2020-07-13 lockepatton/sonipy#2 @faroit
[JOSS] Add unit testing via CI
The package comes without automatic unit tests. Since the application is sonification, the only way to test the functionality of this package is to __listen__ to the outputs. As this is hard to reproduce I would strongly encourage to include a few unit tests that produce tones give a short known x/y sequence.

This issue is part of a [https://github.com/openjournals/joss-reviews/issues/2446](JOSS submission review)
