### 2020-07-13 lockepatton/sonipy#2 @faroit
Update: I didn't saw the test folder being part of the main code folder. I suggest

* moving the test folder to `/tests`
* enable automatic unit testing either via travis, circle.ci or github actions
