### 2020-07-13 lockepatton/sonipy#3 @faroit
[JOSS] replace thinkdsp with librosa.core.tone
the included `thinkdsp.py` file is only used for a single function  - `CosSignal`. I would suggest to copy just this function (and add unit tests for this) or replace this with librosa, a very popular audio framework, that includes a tone generator as well: https://librosa.org/librosa/generated/librosa.core.tone.html

This issue is part of a [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/2446)
