### 2020-07-13 #2446 @faroit
@harpolea @lockepatton To check weather this software meets the JOSS criteria with respect to comparing it to other related packages, I found two other packages.

* https://www.sonification.com.au/sonipy/index.html
* https://github.com/interactive-sonification/pya

Both of them can be used for auditory display/sonification, which is the purpose of this software.
One of them (sonipy) even carries the same name as this submission. Before continuing with here, I guess that it would be recommended to rename this software. Furthermore, it is suggest to includes a detailed comparison to related packages in the paper.
