### 2020-07-13 lockepatton/sonipy#4 @faroit
[JOSS] documentation
From the JOSS review criteria list:

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

the core API of this software is not documented in detail. I would suggest to build a sphinx-based (given suitable docstrings) documentation and make this available on either gh-pages or readthedocs.

This issue is part of a [https://github.com/openjournals/joss-reviews/issues/2446](JOSS submission review)
