### 2020-07-13 lockepatton/sonipy#5 @faroit
[JOSS] add community guidelines
The JOSS review criteria recommends to add community guidelines: for a) Contribute to the software b) Report issues or problems with the software c) Seek support

The first point is covered in the readme but in general, a `contributing.md` document and issue templates would be great additions.

This issue is part of a [https://github.com/openjournals/joss-reviews/issues/2446](JOSS submission review)
