### 2020-07-14 #2446 @david-worrall
Hi,
I am the author of the sonipy mentioned.  I am not averse to, would even encourage, an amalgamation of related projects if that is appropriate, as your "A Perceptually Uniform Sonification Package" subtitle suggests it might. My website mentioned, above, is in need of an update as a result of the publication of my recent Sonification Design book (Springer).  I'd be interested in hearing from you, with an outline of your project: dworrall@colu.edu. Thanks. David
