### 2020-07-15 lockepatton/sonipy#4 @benjaminrose
This is a fairly simple package with minimal entry points for the end user. There may be something between the current README and a full sphinx-based cite. However, since the functions are already documented with doc-strings, a sphinx-based system may still be the easiest.
