### 2020-07-15 #2446 @benjaminrose
I like how simple this package makes it to generate an audio equivalent to a scatter plot. However, it seems to need [a few things](https://github.com/lockepatton/sonipy/issues?q=is%3Aissue+%5BJOSS%5D+) before it can be accepted in JOSS. In addition, I would agree with @faroit that the authors need to determine where this package fits in the broader landscape of sonification software.
