### 2020-07-16 #2446 @harpolea
Thank you @faroit and @benjaminrose for your feedback! As it seems like there may be some substantial changes required before this review can proceed, I am going to pause this review and label it as `pending-major-enhancements` for now whilst they are being dealt with.

@lockepatton, please take your time to address the concerns raised above, and let me know when you are ready for me to un-pause the review.
