+++ sonipy/thinkdsp.py
@@ -462,43 +462,43 @@ class Spectrum(_SpectrumParent):
         return Wave(ys, framerate=self.framerate)
 
 
-class IntegratedSpectrum:
-    """Represents the integral of a spectrum."""
-
-    def __init__(self, cs, fs):
-        """Initializes an integrated spectrum:
-
-        cs: sequence of cumulative amplitudes
-        fs: sequence of frequencies
-        """
-        self.cs = np.asanyarray(cs)
-        self.fs = np.asanyarray(fs)
-
-    def plot_power(self, low=0, high=None, expo=False, **options):
-        """Plots the integrated spectrum.
-
-        low: int index to start at
-        high: int index to end at
-        """
-        cs = self.cs[low:high]
-        fs = self.fs[low:high]
-
-        if expo:
-            cs = np.exp(cs)
-
-        plt.plot(fs, cs, **options)
-
-    def estimate_slope(self, low=1, high=-12000):
-        """Runs linear regression on log cumulative power vs log frequency.
-
-        returns: slope, inter, r2, p, stderr
-        """
-        # print self.fs[low:high]
-        # print self.cs[low:high]
-        x = np.log(self.fs[low:high])
-        y = np.log(self.cs[low:high])
-        t = scipy.stats.linregress(x, y)
-        return t
+# class IntegratedSpectrum:
+#     """Represents the integral of a spectrum."""
+#
+#     def __init__(self, cs, fs):
+#         """Initializes an integrated spectrum:
+#
+#         cs: sequence of cumulative amplitudes
+#         fs: sequence of frequencies
+#         """
+#         self.cs = np.asanyarray(cs)
+#         self.fs = np.asanyarray(fs)
+#
+#     def plot_power(self, low=0, high=None, expo=False, **options):
+#         """Plots the integrated spectrum.
+#
+#         low: int index to start at
+#         high: int index to end at
+#         """
+#         cs = self.cs[low:high]
+#         fs = self.fs[low:high]
+#
+#         if expo:
+#             cs = np.exp(cs)
+#
+#         plt.plot(fs, cs, **options)
+#
+#     def estimate_slope(self, low=1, high=-12000):
+#         """Runs linear regression on log cumulative power vs log frequency.
+#
+#         returns: slope, inter, r2, p, stderr
+#         """
+#         # print self.fs[low:high]
+#         # print self.cs[low:high]
+#         x = np.log(self.fs[low:high])
+#         y = np.log(self.cs[low:high])
+#         t = scipy.stats.linregress(x, y)
+#         return t
 
 
 class Dct(_SpectrumParent):
@@ -1103,17 +1103,17 @@ class Wave:
         audio = Audio(data=self.ys.real, rate=self.framerate)
         return audio
 
-
-def unbias(ys):
-    """Shifts a wave array so it has mean 0.
-
-    ys: wave array
-
-    returns: wave array
-    """
-    return ys - ys.mean()
-
-
+#
+# def unbias(ys):
+#     """Shifts a wave array so it has mean 0.
+#
+#     ys: wave array
+#
+#     returns: wave array
+#     """
+#     return ys - ys.mean()
+#
+#
 def normalize(ys, amp=1.0):
     """Normalizes a wave array so the maximum amplitude is +amp or -amp.
 
@@ -1125,41 +1125,41 @@ def normalize(ys, amp=1.0):
     high, low = abs(max(ys)), abs(min(ys))
     return amp * ys / max(high, low)
 
-
-def shift_right(ys, shift):
-    """Shifts a wave array to the right and zero pads.
-
-    ys: wave array
-    shift: integer shift
-
-    returns: wave array
-    """
[…]
