+++ docs/index.rst
@@ -1,0 +1,20 @@
+.. sonipy documentation master file, created by
+   sphinx-quickstart on Thu Sep 10 10:22:45 2020.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to sonipy's documentation!
+==================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
