+++ setup.cfg
@@ -1,2 +1,8 @@
 [metadata]
 description-file = README.md
+
+[options.extras_require]
+docs =
+    sphinx
+    sphinx-automodapi
+    numpydoc
