+++ .readthedocs.yml
@@ -1,0 +1,12 @@
+version: 2
+
+build:
+image: latest
+
+python:
+version: 3.7
+install:
+    - method: pip
+    path: .
+    extra_requirements:
+        - docs
