+++ .readthedocs.yml
@@ -1,12 +1,15 @@
 version: 2
 
 build:
-image: latest
+  image: latest
 
 python:
-version: 3.7
-install:
+  version: 3.7
+  install:
     - method: pip
-    path: .
-    extra_requirements:
+      path: .
+      extra_requirements:
         - docs
+        - all
+
+formats: []
