+++ docs/index.rst
@@ -3,13 +3,199 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to sonipy's documentation!
-==================================
+ ******************
+ sonipy Documentation
+ ******************
 
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
+ ``sonipy`` : A Scatterplot Sonification Package Documentation
+ =============================================================
 
+ |MIT License| |nbviewer| |Binder|
+
+ A package to turn scatter plots into perceptually uniform sound files
+ for use in science and to make science more accessible.
+
+ This project was developed by `Locke
+ Patton <https://twitter.com/Astro_Locke>`__ and `Prof. Emily
+ Levesque <https://twitter.com/emsque>`__. Click
+ `here <https://twitter.com/Astro_Locke/status/1083510515857408000>`__
+ for a twitter thread explaining the motivation behind this project.
+
+ What does ``sonipy`` do?
+ ------------------------
+
+ Here is an `example
+ sonification <https://twitter.com/Astro_Locke/status/1083510562187751424>`__.
+
+ Our method of sonification takes in scatterplot data and produces audio
+ clips that depict each datapoint as a short sound blip with a y value
+ corresponding to pitch and an x value corresponding to arrival time.
+
+ .. figure:: ./paper/images/Method2.png
+    :alt: sonipy setup
+
+    sonipy setup
+
+ **Each data point has a corresponding short tone called a ``blip``,**
+ with a y value corresponding to its pitch and a x value corresponding to
+ its arrival time. Higher y value data points have higher corresponding
+ blip pitches.
+
+ Installation
+ ------------
+
+ ``sonipy`` is pip-installable from command line, as follows:
+
+ .. code:: bash
+
+    pip install sonipy
+
+ Alternately, you can clone the repository and install it yourself, also
+ in command line:
+
+ .. code:: bash
+
+    git clone https://github.com/lockepatton/sonipy.git
+    cd sonipy
+    python setup.py install
+
+ Example Easy Setup
+ ------------------
+
+ For two arrays of the same length, called x and y, you can sonify them
+ using the following:
+
+ .. code:: python
+
+    from sonipy.sonify import SonifyTool
+
+    Tone = SonifyTool(x, y)
+    Tone.play()
+    Tone.save()
+
+ Extended Setup
+ --------------
+
+ If you would like more fine control of the sonification inputs, you can
+ adjust the underlying arguments as follows. For details about the
+ parameters involved, see the Parameter Inputs section below.
+
+ .. code:: python
+
+    from sonipy.sonify import SonifyTool
+
+    C4 = 261.6 # Hz
+    frequency_args = {
+      'frequency_min' : C4,
+      'frequency_max' : C4*4
+      # 'cents_per_value' : -680,
+      # 'value_min' : 0,
+      # 'value_max' : 1,
+    }
+
+    duration_args = {
+      'time_total' : 2000, # ms
+      # 'time_min' : 100, # ms
+      # 'time_max' : 800, # ms
+    }
+
+    duration_scale = 1. / 2000. # x value / time (ms)
+
+    Tone = SonifyTool(x, y,
+                      frequency_args = frequency_args,
+                      duration_args = duration_args,
+                      # duration_scale = duration_scale,
+                      bliplength=0.5)
+
+    Tone.play()
+    Tone.SaveTone()
+
+ Parameter Inputs
+ ----------------
+
+ Frequency Scale Parameters:
+ ~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+ All frequency parameters are entered inside the frequency_args
[…]
