+++ docs/index.rst
@@ -1,190 +1,185 @@
-.. sonipy documentation master file, created by
-   sphinx-quickstart on Thu Sep 10 10:22:45 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
+******************
+sonipy Documentation
+******************
 
- ******************
- sonipy Documentation
- ******************
+``sonipy`` : A Scatterplot Sonification Package Documentation
+=============================================================
 
- ``sonipy`` : A Scatterplot Sonification Package Documentation
- =============================================================
+|MIT License| |nbviewer| |Binder|
 
- |MIT License| |nbviewer| |Binder|
+A package to turn scatter plots into perceptually uniform sound files
+for use in science and to make science more accessible.
 
- A package to turn scatter plots into perceptually uniform sound files
- for use in science and to make science more accessible.
+This project was developed by `Locke
+Patton <https://twitter.com/Astro_Locke>`__ and `Prof. Emily
+Levesque <https://twitter.com/emsque>`__. Click
+`here <https://twitter.com/Astro_Locke/status/1083510515857408000>`__
+for a twitter thread explaining the motivation behind this project.
 
- This project was developed by `Locke
- Patton <https://twitter.com/Astro_Locke>`__ and `Prof. Emily
- Levesque <https://twitter.com/emsque>`__. Click
- `here <https://twitter.com/Astro_Locke/status/1083510515857408000>`__
- for a twitter thread explaining the motivation behind this project.
+What does ``sonipy`` do?
+------------------------
 
- What does ``sonipy`` do?
- ------------------------
+Here is an `example
+sonification <https://twitter.com/Astro_Locke/status/1083510562187751424>`__.
 
- Here is an `example
- sonification <https://twitter.com/Astro_Locke/status/1083510562187751424>`__.
+Our method of sonification takes in scatterplot data and produces audio
+clips that depict each datapoint as a short sound blip with a y value
+corresponding to pitch and an x value corresponding to arrival time.
 
- Our method of sonification takes in scatterplot data and produces audio
- clips that depict each datapoint as a short sound blip with a y value
- corresponding to pitch and an x value corresponding to arrival time.
+.. figure:: ./paper/images/Method2.png
+  :alt: sonipy setup
 
- .. figure:: ./paper/images/Method2.png
-    :alt: sonipy setup
+  sonipy setup
 
-    sonipy setup
+**Each data point has a corresponding short tone called a ``blip``,**
+with a y value corresponding to its pitch and a x value corresponding to
+its arrival time. Higher y value data points have higher corresponding
+blip pitches.
 
- **Each data point has a corresponding short tone called a ``blip``,**
- with a y value corresponding to its pitch and a x value corresponding to
- its arrival time. Higher y value data points have higher corresponding
- blip pitches.
+Installation
+------------
 
- Installation
- ------------
+``sonipy`` is pip-installable from command line, as follows:
 
- ``sonipy`` is pip-installable from command line, as follows:
+.. code:: bash
 
- .. code:: bash
+  pip install sonipy
 
-    pip install sonipy
+Alternately, you can clone the repository and install it yourself, also
+in command line:
 
- Alternately, you can clone the repository and install it yourself, also
- in command line:
+.. code:: bash
 
- .. code:: bash
+  git clone https://github.com/lockepatton/sonipy.git
+  cd sonipy
+  python setup.py install
 
[…]
