+++ sonipy/scales/__init__.py
@@ -1,0 +1,2 @@
+from .durations import *
+from .frequency import *
