### 2020-09-10 e3a55d5
Adding documentation badge
