### 2020-09-11 2eed772
Updating documentation, including adding image destination
