+++ docs/conf.py
@@ -51,3 +51,15 @@ html_theme = 'alabaster'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+
+
+# -- Specifying theme --------------------------------------------------------
+
+import sphinx_rtd_theme
+
+extensions = [
+    ...
+    "sphinx_rtd_theme",
+]
+
+html_theme = "sphinx_rtd_theme"
