+++ docs/conf.py
@@ -28,7 +28,10 @@ author = 'Locke Patton'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 
-extensions = ['sphinx_automodapi.automodapi','numpydoc']
+import sphinx_rtd_theme
+html_theme = "sphinx_rtd_theme"
+
+extensions = ['sphinx_automodapi.automodapi','numpydoc',"sphinx_rtd_theme"]
 numpydoc_show_class_members = False
 
 # Add any paths that contain templates here, relative to this directory.
@@ -51,15 +54,3 @@ html_theme = 'alabaster'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
-
-
-# -- Specifying theme --------------------------------------------------------
-
-import sphinx_rtd_theme
-
-extensions = [
-    ...
-    "sphinx_rtd_theme",
-]
-
-html_theme = "sphinx_rtd_theme"
