### 2020-09-11 lockepatton/sonipy#4 @lockepatton
Thanks! I've given sonipy a sphinx based readthedocs webpage [here](https://sonipy.readthedocs.io/en/latest/), and linked it on the sonipy homepage. Code for this is hosted in ./docs
