+++ requirements.txt
@@ -1,0 +1,6 @@
+numpy>=1.16
+matplotlib
+seaborn
+pandas
+scipy
+pydub
