### 2020-09-11 d79f080
Using pytest, and specifying install in .travis.yml
