+++ .travis.yml
@@ -1,1 +1,7 @@
 language: python
+python:
+  - "3.6"
+  - "3.7"
+install:
+  - pip install .
+script: pytest
