### 2020-09-11 c997cdc
The Tone.play() function cannot be tested on travis, as not hosting in IPython environment. Adding try, except for such a case.
