### 2020-09-11 6a11cf9
Added Build Status to README; removed requirements.txt
