+++ requirements.txt
@@ -1,6 +1,0 @@
-numpy>=1.16
-matplotlib
-seaborn
-pandas
-scipy
-pydub
