### 2020-09-11 f30874e
Testing IPython in requirements to make play() work in testing
