+++ tests/test_sonify.py
@@ -138,15 +138,7 @@ class TestMultiTone(TestCase):
     def test_play(self):
         # check to make sure .play(), the wrapper of ThinkDSP .make_audio(), runs
         scale, Tone = createTone(x_lin, y_lin)
-
-        try:
-            from IPython.display import Audio
-            Tone.play()
-        except:
-            import warnings
-            warnings.warn("Can't import Audio from IPython.display; most likely not using jupyter noteebok.")
-            pass
-
+        Tone.play()
 
     def test_raises_exception_if_starttimes_and_values_not_equal_length(self):
         times, values = x_lin[:4], y_lin
