### 2020-09-11 lockepatton/sonipy#2 @lockepatton
I've moved and connected the ./tests files to travis - and the passing badge is now in the README.md. 🎉
