### 2020-10-07 dad1850
Added contributing.md
