### 2020-10-07 1670994
contributing.md minor additions and format edits
