### 2020-10-07 lockepatton/sonipy#5 @lockepatton
Thanks for this idea - I've implemented it in the most recent commit. Let me know if you have anything of a similar vein in mind; for now, I'm closing the issue here.
