+++ sonipy/thinkdsp.py
@@ -462,45 +462,6 @@ class Spectrum(_SpectrumParent):
         return Wave(ys, framerate=self.framerate)
 
 
-# class IntegratedSpectrum:
-#     """Represents the integral of a spectrum."""
-#
-#     def __init__(self, cs, fs):
-#         """Initializes an integrated spectrum:
-#
-#         cs: sequence of cumulative amplitudes
-#         fs: sequence of frequencies
-#         """
-#         self.cs = np.asanyarray(cs)
-#         self.fs = np.asanyarray(fs)
-#
-#     def plot_power(self, low=0, high=None, expo=False, **options):
-#         """Plots the integrated spectrum.
-#
-#         low: int index to start at
-#         high: int index to end at
-#         """
-#         cs = self.cs[low:high]
-#         fs = self.fs[low:high]
-#
-#         if expo:
-#             cs = np.exp(cs)
-#
-#         plt.plot(fs, cs, **options)
-#
-#     def estimate_slope(self, low=1, high=-12000):
-#         """Runs linear regression on log cumulative power vs log frequency.
-#
-#         returns: slope, inter, r2, p, stderr
-#         """
-#         # print self.fs[low:high]
-#         # print self.cs[low:high]
-#         x = np.log(self.fs[low:high])
-#         y = np.log(self.cs[low:high])
-#         t = scipy.stats.linregress(x, y)
-#         return t
-
-
 class Dct(_SpectrumParent):
     """Represents the spectrum of a signal using discrete cosine transform."""
 
@@ -1103,17 +1064,6 @@ class Wave:
         audio = Audio(data=self.ys.real, rate=self.framerate)
         return audio
 
-#
-# def unbias(ys):
-#     """Shifts a wave array so it has mean 0.
-#
-#     ys: wave array
-#
-#     returns: wave array
-#     """
-#     return ys - ys.mean()
-#
-#
 def normalize(ys, amp=1.0):
     """Normalizes a wave array so the maximum amplitude is +amp or -amp.
 
@@ -1125,41 +1075,6 @@ def normalize(ys, amp=1.0):
     high, low = abs(max(ys)), abs(min(ys))
     return amp * ys / max(high, low)
 
-#
-# def shift_right(ys, shift):
-#     """Shifts a wave array to the right and zero pads.
-#
-#     ys: wave array
-#     shift: integer shift
-#
-#     returns: wave array
-#     """
-#     res = np.zeros(len(ys) + shift)
-#     res[shift:] = ys
-#     return res
-#
-#
-# def shift_left(ys, shift):
-#     """Shifts a wave array to the left.
-#
-#     ys: wave array
-#     shift: integer shift
-#
-#     returns: wave array
-#     """
-#     return ys[shift:]
-#
-#
-# def truncate(ys, n):
-#     """Trims a wave array to the given length.
-#
-#     ys: wave array
-#     n: integer length
-#
-#     returns: wave array
-#     """
-#     return ys[:n]
-#
 
 def quantize(ys, bound, dtype):
     """Maps the waveform to quanta.
@@ -1176,36 +1091,6 @@ def quantize(ys, bound, dtype):
 
     zs = (ys * bound).astype(dtype)
     return zs
-
-
-# def apodize(ys, framerate, denom=20, duration=0.1):
-#     """Tapers the amplitude at the beginning and end of the signal.
-#
-#     Tapers either the given duration of time or the given
-#     fraction of the total duration, whichever is less.
-#
-#     ys: wave array
-#     framerate: int frames per second
-#     denom: float fraction of the segment to taper
-#     duration: float duration of the taper in seconds
-#
-#     returns: wave array
-#     """
-#     # a fixed fraction of the segment
-#     n = len(ys)
-#     k1 = n // denom
-#
[…]
