### 2020-11-05 lockepatton/sonipy#3 @lockepatton
Hi @faroit,

So sorry - this has been on the back burner, but I'm ready to make this a priority these next weeks.

I looked into the package you referenced, and found it couldn't do some of the minor odd specifics I needed. Switching to librosa would require more of a major rehaul of the code that feels necessary given our scope.

Looking into thinkdsp.py, there are some major class dependencies that the code uses in order to run the simple Sinusoid object, but I've pulled together some basic tests that have now been committed in ./tests/test_thinkdsp.py.

For your reference, I've commented out the portions of thinkdsp.py that are not strictly necessary. While the tests don't cover 100% of the processes, I think my tests cover the basics. Let me know if this is enough to fix this issue.

Cheers, Locke
