### 2020-11-05 #2446 @lockepatton
Hi everyone,

Apologies for being slow in addressing these changes, and thanks for the reminder @arfon. I've made a commit on the last remaining issue. If the changes I've committed there are sufficient, we can unpause the review.

I'm going to make this a priority this week and the next so we can finish this submission. Thanks.
