### 2020-11-05 #2446 @arfon
Thanks @lockepatton - for clarity could you link to the _"commit on the last remaining issue"_ here so we can direct the reviewers back to this?

Also, both reviewers asked for clarification on the relationship with other related community efforts. A writeup of any outcomes/updates here from your perspective would be important before we un-pause this submission.
