### 2020-11-06 lockepatton/sonipy#3 @faroit
> I looked into the package you referenced, and found it couldn't do some of the minor odd specifics I needed. Switching to librosa would require more of a major rehaul of the code that feels necessary given our scope.

Curious about the specifics that you need. I am sure @bmcfee would love to hear about them and you could think of adding this to librosa via PR, which would allow more users to get access to these functionalities.

If this is too specific, I am okay with removing the unnecessary code form thinkdsp.py
