### 2020-11-18 lockepatton/sonipy#3 @lockepatton
Hi @faroit - thanks for the reply. Looking back over this, I think removing the unnecessary code is the best way to go. I've already commented the unnecessary code out; let me know if you would rather I delete it entirely or if leaving it commented (for possible future updates) is okay. At this point, let me know if you need anything else from me to complete revisions for this submission. Thanks!
