### 2020-11-18 #2446 @lockepatton
Hi @arfon!
[Here](https://github.com/lockepatton/sonipy/issues/3) is the outstanding issue I was referencing.

**Re the clarification on the relationship with other community efforts:**

This sonipy code is specifically designed to create an open-access scientifically useful method to listen to data, with accuracy and use on par with reading plots visually. While many sonification tools exist, this was specifically designed in collaboration with the University of Washington Hearing and Speech Sciences to guarantee that a linear increase in y value will correspond to a perceptually uniform increase in perceived “pitch”. This means that while frequency varies non linearly, the user is listening to data in a perceptually uniform space - a linear plot sounds like a linear sweep in pitch. This technique allows us to probe smaller difference in y values than we can discern visually. We’ve also found we can hear periodic details in data that are not easily visually discerned.

Thanks to the nature of human hearing, we can audibly discern subsequent pitch differences of 10 cents (a logarithmic measure of pitch interval). On a y scale ranging from 0 to 10, that corresponds to hearing variations as small as dy~0.03 - a number which rivals our visual perceptions of scatter plot detail. This simultaneous depth and range makes pitch-varied audio an incredibly powerful and accessible tool for understanding nuances in data. This approach also opens up science and citizen science to participants who are visually impaired, and empowers blind and visually impaired (BVI) individuals to explore their own data.

In fact, this code was specifically designed for this use case. Already it’s been used by students at Perkin’s School for the Blind, a semester Ohio State Astronomy Program for BVI high school students, and is the foundation for an NSF-funded TransientZoo project, a citizen science program that will allow participants, including BVI individuals, to classify astronomical supernova lightcurves using sound. For more details, [here](https://github.com/lockepatton/sonipy/blob/master/paper/paper.md) is a link to my paper where I go into the TransientZoo case study.

**re: the state of sonification tools**

The two codes that you linked above - @faroit  - are both very interesting but distinct sonification cases. In emailing with @david-worrall, the author of the other code “[sonipy](https://www.sonification.com.au/sonipy/index.html)”, it’s clear that their code is more an awesome cumulation of interface codes. While a potentially cool idea in the future to connect this sonipy with their infrastructure to connect more with the field of sonification, this code is most useful and accessible to BVI students as already designed and implemented. It’s a tool designed to be plugged into various science disciplines (in fields otherwise unfamiliar with sonification) and thus fits well without a complete sonification library.

The second code [interactive-sonification](https://github.com/interactive-sonification/pya) is awesome and designed for interactive audio processing. Our code is specifically designed non-interactively; while it uses audio processing, it uses [thinkDSP](https://github.com/AllenDowney/ThinkDSP) as the audio processing tool. Our code does the work of designing a perceptually uniform sonification for accurate use in listening to science plots. We specially built this tool as there was no other sonification that filled this use.

Since I last updated everyone, I've closed the last issue, so if everyone agrees (especially @faroit as we were resolving the last issue [here]((https://github.com/lockepatton/sonipy/issues/3)), I believe we can unpause this review. Let me know if you need anything else from me in the meantime.
