### 2020-11-19 lockepatton/sonipy#3 @lockepatton
Planning on closing this issue, but feel free to reopen if I misread your comment @faroit.
