### 2020-11-22 #2446 @arfon
Got it, thanks for the detailed response @lockepatton.

@faroit, @benjaminrose - I realize it's been quite some time since you both agreed to this review but, provided you find @lockepatton's responses above satisfactory, would you both be willing to pick up yours reviews sometime soonish?

Many thanks!
