### 2020-12-03 #2446 @benjaminrose
I do not think this is project is currently at an acceptable state for JOSS, however, it is close. The main concern is over the substantialness of the scholarly effort. @lockepatton, a strong statement of need will alleviate this concern. Looking at the JOSS guidelines, I think this project "is likely to be cited by your peer group" if the need for this specific project is well addressed.

Before accepting, I see the following required changes.

- [ ] Update the documentation and paper to more clearly address why this project is needed over other projects already available.
- [ ] The above should also simultaneously address the other aspects missing from the paper (i.e. state of the field).
- [ ] Update the documentation to clearly (likely with links) address how all three third parties participants can interact with the project: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. Currently, all three of these are covered by the generic "Have an issue with your operating system? Let us know by opening an issue!"
- [ ] Ensure that @faroit is also satisfied with the state of the project.
