### 2020-12-03 #2446 @arfon
Thanks for your update @benjaminrose. These sound like good recommendations for @lockepatton to address.

>  Ensure that @faroit is also satisfied with the state of the project.

@faroit - how are you getting along here? I see there are still a number of unchecked items in your review checklist.
