### 2020-12-03 #2446 @david-worrall
Hello all,

Perceptual uniformity is a complex interdisciplinary research field. It
is a big claim in of this project, not yet substantiated. It would bode
well for the credibility of the project itself to temper that claim, as
there is more work to do towards perceptual uniformity that is beyond
the use of (some) psychophysical mapping parameters.
For example, see the work of  Roddy et al. on cognitive metaphors, and
Friberg et al. (with /Director Musices/ for example) - as overviewed in
Ch 4 of my book.mThis is where the perceptual research is largely
developing in the field, including in my own /Sonipy /software framework.

=== On the use of the name /Sonipy ===/
My extensive and still active project of that name has been evolving for
more than 10 years. It does not make sense to confuse the
research/development/user space with a new project of the same name. The
'noise' so generated is at best inconvenient, especially as there are
other viable alternatives; /Pysonic /for example seems available.

I wish the project well and, as I indicated in correspondence, am not
averse to a friendly merger, once this project is past the beta stage.

---

Dr David Worrall
Professor, Audio Arts and Acoustics
Columbia College Chicago
33 E. Ida B. Wells Drive Room 601A
Chicago, ILLINOIS, USA 60605

<https://en.wikipedia.org/wiki/David_Worrall_(composer)>Wikipedia<https://en.wikipedia.org/wiki/David_Worrall_(composer)>Instagram<https://www.instagram.com/david.r.worrall/>Facebook<https://www.facebook.com/david.worrall.5872/>Soundcloud<https://soundcloud.com/david-worrall-8849455>Youtube<https://www.youtube.com/playlist?list=PL9vjIQ8ca6oOnhOs1fMq1DqhoaxXV-9i0>
Personal research/creative practice website:
avatar.com.au<http://avatar.com.au/>
New Organised Sound issue (25/1): Computation in the Sonic
Arts<https://www.cambridge.org/core/journals/organised-sound/volume/C868728E57C2FEF58217164D7F862577>
New Book: Sonification Design: From data to intelligible
soundfields<https://www.springer.com/gp/book/9783030014964>(Springer)
(also available through
Amazon<https://www.amazon.com/Sonification-Design-intelligible-soundfields-Human-Computer/dp/3030014967/ref=sr_1_fkmrnull_1?keywords=worrall+sonification&qid=1553803176&s=gateway&sr=8-1-fkmrnull>)
