### 2020-12-04 lockepatton/sonipy#3 @faroit
> I think removing the unnecessary code is the best way to go. I've already commented the unnecessary code out; let me know if you would rather I delete it entirely or if leaving it commented (for possible future updates) is okay. At this point, let me know if you need anything else from me to complete revisions for this submission. Thanks!

i just had a quick look and I still think that there is a lot of code in `thinkdsp.py` that is not used from within sonipy, I would recommend to strip down everything except the parts that are useful for running `CosSignal`
