### 2020-12-04 #2446 @faroit
1. I fully support the comments and suggestions made by @benjaminrose especially that the statements with respect to state-of-the-art should be added to the paper.

2. I support the view of @david-worrall that the same name for a similar project is confusing (even though david's package is neither on github nor on pypi). Given that merging the projects will probably not happen soon, it is the best for this submission to be renamed before being accepted.

3. I still struggle with checking `Substantial scholarly effort`. The reason is that the submission has barely more than 300LOC when ignoring the copied `thinkdsp.py` module as mentioned in https://github.com/lockepatton/sonipy/issues/3.  I am okay with a very lightweight package being accepted but I would suggest to at least revisit this issue, refactor the code and preferably include librosa as a dependency.
