### 2020-12-05 #2446 @arfon
> I still struggle with checking Substantial scholarly effort. The reason is that the submission has barely more than 300LOC when ignoring the copied thinkdsp.py module as mentioned in lockepatton/sonipy#3. I am okay with a very lightweight package being accepted but I would suggest to at least revisit this issue, refactor the code and preferably include librosa as a dependency.

Thanks for flagging this @faroit. I had not realized that `thinkdsp.py` was essentially a dependency here, and not functionality implemented by the author.

@benjaminrose - I notice that you also are yet to check this box off. I'd appreciate your thoughts on this point too (our submission guidelines are here: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
