### 2020-12-07 #2446 @benjaminrose
@arfon, this is a small project and the scholarly effort is near JOSS’s minimum requirement. In general, I think that science benefits from allowing more works to be published rather than fewer.

This specific case misses most of the examples in the JOSS documentation. However, I feel like this project could have a major impact on the field. In the study of odd astrophysical transients, a field that will see exponential growth in the next decade, light curve comparisons are the main scientific figure. This project has the potential to add audio “light curves” with fewer lines of code than you typically use to reformat a default `matplotlib` figure. The addition of audio light curves might allow for more inclusive science, viral public outreach, as well as a better scientific understanding of the minor differences in objects.

However, my expectations are not given, and another project may already do this sufficiently well. So a strong statement of need, description of the state-of-the-art, and explanation of how this project is unique is vital to show that though small, it has the potential impact sufficient enough to claim substantial scholarly effort.
