### 2021-01-19 #2446 @benjaminrose
I regret to inform @lockepatton and others that it appears Space Telescope Science Institute has also created a sonification package, [Astronify](https://astronify.readthedocs.io/en/latest/). This makes the need for `sonipy` even more difficult to explain.
