### 2021-01-20 #2446 @lockepatton
Thanks for bringing up astronify: this STScI code is, in fact, actually based on our package. We’ve been in touch with the creators (Clara and Scott) over email and video since March of 2020 (when they reached out to us following our presentation of this work at last year’s American Astronomical Society meeting) and have been advising them on the basics of sonification and audio perception since they first began their project (though email, video chat, and regular meetings of the World Sonification Chats). Indeed, we’re mainly just disappointed that they’ve now released and begun promoting their code publicly without consulting or citing us in any way, particularly as their code is directly derived from our work and they were aware that we were in the process of publishing our code as peer-reviewed work in JOSS.

We’d be happy to add a mention of astronify to our paper as yet another example of the impact our code is having on the field of data sonification in astronomy. We’ll also be reminding the astronify creators that they should properly acknowledge our contribution via a JOSS citation.

Apart from adding this mention of astronify, we are almost done incorporating our state-of-the-field summary, making the other minor changes you suggested to the paper, and adding the requested additional documentation; we anticipate that this will be finished by Friday at the latest. Once these changes are made we hope that our package will be considered ready for publication. Please let us know if you need any additional changes to the paper, or if you’d like further clarification on where astronify fits in the sonification landscape as a project derived from our code.
