### 2021-01-23 4d5957c
changes from referee paper comments
