+++ paper/paper.md
@@ -1,15 +1,15 @@
 ---
-title: '`Sonipy`: A Perceptually Uniform Sonification Package'
+title: 'sonipy: A Perceptually Uniform Sonification Package'
 tags:
   - Python
   - Sonification
   - Astronomy
   - Supernova
 authors:
-  - name: Locke Patton^[locke.patton\@cfa.harvard.edu]
+  - name: Locke Patton^[locke.patton@cfa.harvard.edu]
     orcid: 0000-0002-7640-236X
     affiliation: "1, 2"
-  - name: Emily Levesque^[emsque\@uw.edu]
+  - name: Emily Levesque^[emsque@uw.edu]
     affiliation: 1
 affiliations:
  - name: University of Washington, Department of Astronomy,  Seattle, WA 98195 USA
@@ -20,11 +20,13 @@ date: 1 July 2020
 bibliography: paper.bib
 ---
 
+![Example sonification case: an exploding star's change in brightness is plotted against time. Each data point corresponds to a tone blip at a frequency specified by its y value and a time specified by its x value. As the sound file plays, it scans the plot left to right, with the brightest moments of the exploding star reaching pitches of 3 times middle C on the piano and the tail of the cooling supernovae remnant dropping into lower audible pitches.](./images/Picture1-nobkgd.png)
+
 # Introduction
 
-`Sonipy` moves beyond visual analyses by sonifying scatter-plot data, producing audio files that depict variations in y as perceptually uniform changes in pitch. Short tones - called blips - are sounded in time at intervals corresponding to x values. We built this code to open up science to participants who are visually impaired, creating scientifically useful and distinguishable sonifications of scatter-plot data.
+`Pysonic` moves beyond visual analyses by sonifying scatter-plot data, producing audio files that depict variations in y as perceptually uniform changes in pitch. Short tones - called blips - are sounded in time at intervals corresponding to x values.
 
-![Example sonification case: an exploding star's change in brightness is plotted against time. Each data point corresponds to a tone blip at a frequency specified by its y value and a time specified by its x value. As the sound file plays, it scans the plot left to right, with the brightest moments of the exploding star reaching pitches of 3 times middle C on the piano and the tail of the cooling supernovae remnant dropping into lower audible pitches.](./images/Picture1-nobkgd.png)
+The addition of this audio sonification to a scientist's toolset has the intention of creating more inclusive and accessible science, new attention grabbing public outreach opportunities, and an easier self-consistent scientific understanding of minor differences in data over large scales.
 
 ## Understanding pitch
 
@@ -32,19 +34,31 @@ The cent is a logarithmic unit of measure for pitch intervals where $n \approx 3
 
 ## Human Pitch Sensitivity
 
-The average person is capable of discerning independent subsequent pitches with a difference of ~10 cents [@Kollmeier:2008]. The human ear is most sensitive to frequencies between ~500-4000 Hz, similar to the range of a standard piano.
+The average person is capable of discerning independent subsequent pitches with a difference of ~10 cents [@Kollmeier:2008]. The human ear is most sensitive to frequencies between $\sim$500-4000 Hz, similar to the range of a standard piano.
 
-With these parameters, xy scatter-plot data can be translated into audio files that map y values to specific pitch frequencies, with the minimum discernible $\Delta y$ corresponding to a 10 cent pitch difference.
+With these parameters, xy scatterplot data can be translated into audio files that map y values to specific pitch frequencies, with the minimum discernible $\Delta y$ corresponding to a 10 cent pitch difference.
 
-# The Case for Sonification: Why sonify lightcurves?
+# Scientific Need
 
[…]
