### 2021-01-23 #2446 @lockepatton
Our edits to the paper itself are now complete! The requested changes to the code documentation will also be finished shortly, and we expect to submit those by Monday.
