### 2021-04-07 158de04
Import gcd from math instead of fractions for Python 3.9 compatibility.
