### 2021-04-12 a1fd86d
Testing python 3.4-3.9 via travis
