+++ .travis.yml
@@ -2,6 +2,8 @@ language: python
 python:
   - "3.6"
   - "3.7"
+  - "3.8"
+  - "3.9"
 install:
   - pip install .
 script: pytest
