### 2021-04-12 98c866f
Changing matplotlib requirement for py 3.7
