+++ setup.py
@@ -21,4 +21,4 @@ setuptools.setup(
          'License :: OSI Approved :: MIT License',
          'Operating System :: OS Independent',
      ],
-     install_requires=['numpy>=1.16.4', 'matplotlib', 'seaborn', 'pandas', 'scipy', 'pydub','IPython'])
+     install_requires=['numpy>=1.16.5', 'matplotlib', 'seaborn', 'pandas', 'scipy', 'pydub','IPython'])
