+++ README.md
@@ -1,8 +1,8 @@
-pysonicpysonic# `pysonic` : A Scatterplot Sonification Package
+sonipysonipysonipy# `sonipy` : A Scatterplot Sonification Package
 
-[![MIT License](https://img.shields.io/apm/l/atomic-design-ui.svg?)](https://github.com/lockepatton/pysonic/blob/master/LICENSE.txt)
-[![nbviewer](https://img.shields.io/badge/jupyter%20notebooks-nbviewer-blue)](https://nbviewer.jupyter.org/github/lockepatton/pysonic/blob/master/demos/Tutorial.ipynb)
-[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/lockepatton/pysonic/master?filepath=demos%2FTutorial.ipynb)
+[![MIT License](https://img.shields.io/apm/l/atomic-design-ui.svg?)](https://github.com/lockepatton/sonipy/blob/master/LICENSE.txt)
+[![nbviewer](https://img.shields.io/badge/jupyter%20notebooks-nbviewer-blue)](https://nbviewer.jupyter.org/github/lockepatton/sonipy/blob/master/demos/Tutorial.ipynb)
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/lockepatton/sonipy/master?filepath=demos%2FTutorial.ipynb)
 [![Documentation Status](https://readthedocs.org/projects/sonipy/badge/?version=latest)](https://sonipy.readthedocs.io/en/latest/?badge=latest)
 [![Build Status](https://api.travis-ci.org/lockepatton/sonipy.png?branch=master)](https://travis-ci.org/github/lockepatton/sonipy)
 
@@ -10,29 +10,29 @@ A package to turn scatter plots into perceptually uniform sound files for use in
 
 This project was developed by [Locke Patton](https://twitter.com/Astro_Locke) and [Prof. Emily Levesque](https://twitter.com/emsque). Click [here](https://twitter.com/Astro_Locke/status/1083510515857408000) for a twitter thread explaining the motivation behind this project.
 
-## What does `pysonic` do?
+## What does `sonipy` do?
 
 Here is an [example sonification](https://twitter.com/Astro_Locke/status/1083510562187751424).
 
 Our method of sonification takes in scatterplot data and produces audio clips that depict each datapoint as a short sound blip with a y value corresponding to pitch and an x value corresponding to arrival time.
 
-![pysonic setup](./paper/images/Method2.png)
+![sonipy setup](./paper/images/Method2.png)
 
 **Each data point has a corresponding short tone called a `blip`,** with a y value corresponding to its pitch and a x value corresponding to its arrival time. Higher y value data points have higher corresponding blip pitches.
 
 ## Installation
 
-`pysonic` is pip-installable from command line, as follows:
+`sonipy` is pip-installable from command line, as follows:
 
 ``` bash
-pip install pysonic
+pip install sonipy
 ```
 
 Alternately, you can clone the repository and install it yourself, also in command line:
 
 ``` bash
-git clone https://github.com/lockepatton/pysonic.git
-cd pysonic
+git clone https://github.com/lockepatton/sonipy.git
+cd sonipy
 python setup.py install
 ```
 
@@ -41,7 +41,7 @@ python setup.py install
 For two arrays of the same length, called x and y, you can sonify them using the following:
 
 ``` Python
-from pysonic.sonify import SonifyTool
+from sonipy.sonify import SonifyTool
 
 Tone = SonifyTool(x, y)
 Tone.play()
@@ -53,7 +53,7 @@ Tone.save()
[…]
