### 2021-04-13 #2446 @lockepatton
Hi @afron,

Apologies for the delay - things are going well. I believe I've completed most of the requests you asked for listed above. Two are outstanding:

- This code still needs a name. I've tried the name you suggested - pysonic - but it has already been populated on github and pip. No luck in finding another better alternative that has not already been similarly populated, although I did research upwards of thirty ideas. I would hope that keeping sonipy would be acceptable given this. It would be hard to change it but if necessary for submission I can work to complete that.

- To your third point, in using librosa as a library - including librosa would be a large coding project, as a lot of the fundamental design of my objects is based on thinkdsp.py. My original choice of audio tool was based upon testing the sonification tools available to me when the code was under development. This was the final best choice at the time. We tested upwards of three different sonification codes and found this to produce the best sound. As changing to librosa would not fundamentally change any aspect of the code for the user, I'm hesitant to make such a change and nervous the sound quality might differ.

I'm happy to talk in real time if that would be helpful.

Thanks for your suggestions and patience. Pandemic year is not the best year to publish papers apparently.
