### 2021-04-26 874ccd3
Simplified thinkdsp.py to basic commands, without extraneous functions/code; added note to tutorial.ipynb
