+++ sonipy/thinkdsp.py
@@ -72,7 +72,7 @@ class WavFileWriter:
         self.fp.setnchannels(self.nchannels)
         self.fp.setsampwidth(self.sampwidth)
         self.fp.setframerate(self.framerate)
-
+#
     def write(self, wave):
         """Writes a wave.
 
@@ -90,66 +90,6 @@ class WavFileWriter:
             self.write(rest(duration))
 
         self.fp.close()
-
-
-def read_wave(filename="sound.wav"):
-    """Reads a wave file.
-
-    filename: string
-
-    returns: Wave
-    """
-    fp = open_wave(filename, "r")
-
-    nchannels = fp.getnchannels()
-    nframes = fp.getnframes()
-    sampwidth = fp.getsampwidth()
-    framerate = fp.getframerate()
-
-    z_str = fp.readframes(nframes)
-
-    fp.close()
-
-    dtype_map = {1: np.int8, 2: np.int16, 3: "special", 4: np.int32}
-    if sampwidth not in dtype_map:
-        raise ValueError("sampwidth %d unknown" % sampwidth)
-
-    if sampwidth == 3:
-        xs = np.fromstring(z_str, dtype=np.int8).astype(np.int32)
-        ys = (xs[2::3] * 256 + xs[1::3]) * 256 + xs[0::3]
-    else:
-        ys = np.fromstring(z_str, dtype=dtype_map[sampwidth])
-
-    # if it's in stereo, just pull out the first channel
-    if nchannels == 2:
-        ys = ys[::2]
-
-    # ts = np.arange(len(ys)) / framerate
-    wave = Wave(ys, framerate=framerate)
-    wave.normalize()
-    return wave
-
-
-def read_wave_with_scipy(filename):
-    """Reads a wave file.
-
-    filename: string
-
-    returns: Wave
-    """
-    # TODO: Check back later and see if this works on 24-bit data,
-    # and runs without throwing warnings.
-    framerate, ys = wavfile.read(filename)
-
-    # if it's in stereo, just pull out the first channel
-    if ys.ndim == 2:
-        ys = ys[:, 0]
-
-    # ts = np.arange(len(ys)) / framerate
-    wave = Wave(ys, framerate=framerate)
-    wave.normalize()
-    return wave
-
 
 def play_wave(filename="sound.wav", player="aplay"):
     """Plays a wave file.
@@ -209,61 +149,6 @@ class _SpectrumParent:
         Returns: new Spectrum
         """
         return copy.deepcopy(self)
-
-    def max_diff(self, other):
-        """Computes the maximum absolute difference between spectra.
-
-        other: Spectrum
-
-        returns: float
-        """
-        assert self.framerate == other.framerate
-        assert len(self) == len(other)
-
-        hs = self.hs - other.hs
-        return np.max(np.abs(hs))
-
-    def ratio(self, denom, thresh=1, val=0):
-        """The ratio of two spectrums.
-
-        denom: Spectrum
-        thresh: values smaller than this are replaced
-        val: with this value
-
-        returns: new Wave
-        """
-        ratio_spectrum = self.copy()
-        ratio_spectrum.hs /= denom.hs
-        ratio_spectrum.hs[denom.amps < thresh] = val
-        return ratio_spectrum
-
-    def invert(self):
-        """Inverts this spectrum/filter.
-
-        returns: new Wave
-        """
-        inverse = self.copy()
-        inverse.hs = 1 / inverse.hs
-        return inverse
-
-    @property
-    def freq_res(self):
-        return self.framerate / 2 / (len(self.fs) - 1)
-
-    def render_full(self, high=None):
-        """Extracts amps and fs from a full spectrum.
-
-        high: cutoff frequency
-
[…]
