### 2021-05-03 #2446 @lockepatton
Hi @arfon  - following up on this request again. The two questions outstanding are on the use of sonipy and the use of thinkdsp.py. Let me know if there are other open tasks, but for now I'm awaiting your reply on the best way to move forward.
