### 2021-05-04 #2446 @arfon
> Hi @arfon - following up on this request again. The two questions outstanding are on the use of sonipy and the use of thinkdsp.py. Let me know if there are other open tasks, but for now I'm awaiting your reply on the best way to move forward.

Sorry for the delay @lockepatton.

Given that `sonipy` already exists, I think it would be appropriate to find a different name. Could something like `sonicpy`, `py-sonic`, or `python-sonic` work?

> To your third point, in using librosa as a library - including librosa would be a large coding project, as a lot of the fundamental design of my objects is based on thinkdsp.py. My original choice of audio tool was based upon testing the sonification tools available to me when the code was under development. This was the final best choice at the time. We tested upwards of three different sonification codes and found this to produce the best sound. As changing to librosa would not fundamentally change any aspect of the code for the user, I'm hesitant to make such a change and nervous the sound quality might differ.

While I think @faroit's [recommendation](https://github.com/openjournals/joss-reviews/issues/2446#issuecomment-739066814) is a good one, at this point, I think it's OK to move forward with using `thinkdsp.py` as the dependency here rather than ask @lockepatton to embark on a significant refactor.
