### 2021-05-17 #2446 @lockepatton
Hi @arfon and @faroit!

Perfect. I really agree that librosa could be an awesome eventual upgrade - just not at this moment before publication.

As for the name!  ```sonicpy``` is not successfully google-able due to the sonic franchise and already has software with that name.  ```pysonic``` similarly already has a collection of different softwares with that name. May I suggest ``soni-py```? That way the publications we already have with sonipy in the title do not become null and void, but the name is slightly different to avoid confusion with the other ```soniny```. If this sounds good to you, let us know and we'll update all documentation accordingly, as well as update pypl where it is already published under sonipy.
