### 2021-05-17 #2446 @faroit
> May I suggest soni-py? That way the publications we already have with sonipy in the title do not become null and void, but the name is slightly different to avoid confusion with the other soniny.

sounds good to me. Just be aware of the python limitations for hyphens vs. underscores. pypi supports both, I think...
