### 2021-05-17 #2446 @arfon
> If this sounds good to you, let us know and we'll update all documentation accordingly, as well as update pypl where it is already published under sonipy.

Also sounds good to me 👍
