### 2021-05-17 #2446 @lockepatton
Alight - so I started working on the name change and ran into a hiccup. [This](https://stackoverflow.com/questions/54597212/using-hyphen-dash-in-python-repository-name-and-package-name) stack overflow describes this very well.
Basiclaly, I can superficially change the name to soni-py in documentation, but need to leave the python interactable places as sonipy -- these include pypi and the home name of my directory so that the following import does not result in a syntax error: ```from sonipy.sonify import SonifyTool``` . Does this seem like an acceptable way to go forward?
