### 2021-05-17 63dde31
Tutorial.ipynb update and name update in contributions.md and tutorial
