### 2021-05-18 #2446 @david-worrall
There are (still) two elephants in this room.

_1. The package claim:_ "A Perceptually Uniform Sonification Package"
Unless a /huge/ amount of work has gone on in the interim between now
and when I last looked,
"A Perceptually Uniform Sonification Package" oversells the generality
of it:

    - The "perceptions" are restricted to that of frequency as pitch is
    not a revelation. Over the history of the field, the number of
    sonifications that don't treat frequency like that could be counted
    on the fingers of one hand. Even then, it only applies the (linear,
    log) mapping in its simplest form, without taking into account
    Fletcher-Munson and related psychophysical mappings and other
    cross-modal sensory affects.  See
    https://en.wikipedia.org/wiki/Equal-loudness_contour.
    - There there seems no model of a perceptually uniform timbre space,
    on which there is a considerable literature.
    -There there seems no model of a perceptually uniform temporal
    space, let alone models of pulse, meter, etc

_2. The package name: sonipy_

Surely, given the more than a decade of use of this name for another
purpose (see, for example, overviews in Chapters 5 and 6 of Sonification
Design: From data to intelligible soundfields
<https://www.springer.com/gp/book/9783030014964>(Springer) it would be
better, and more intellectually honest, to not muddy those waters and
just choose another name. If for no other reason, the confusion
generated by not doing so is likely to negatively impact on the
reception of Package #2446 and the perception of the value of that
project as a whole.
