### 2021-05-18 #2446 @faroit
@arfon @david-worrall thanks for your clarifications – especially to _1_.  However, the paper subtitle has already been changed to _Pysonic: A Pitch-Based Sonification Package_, so I assume this is a more suitable subtitle. @lockepatton please change it also in the issue title etc.

Regarding the naming issue, as I pointed out earlier, we should encourage authors to respect existing software projects, including its name. @lockepatton to not speed up the reviewing process I suggest to find a new name which is less close to the existing package by David Worral.
