### 2021-05-19 #2446 @arfon
> @arfon @david-worrall thanks for your clarifications – especially to 1. However, the paper subtitle has already been changed to Pysonic: A Pitch-Based Sonification Package, so I assume this is a more suitable subtitle. @lockepatton please change it also in the issue title etc.

Right, thanks for pointing this out @faroit. I've updated the review thread to match the paper title.

On the naming issue: This is a challenging situation especially given that it sounds like there are already papers published that link to this tool under the name `sonipy`. Ultimately I don't think JOSS can (or should) rule on package naming but I would like to consult with the broader JOSS editorial team on this point
