### 2021-05-19 #2446 @faroit
> On the naming issue: This is a challenging situation especially given that it sounds like there are already papers published that link to this tool under the name sonipy. Ultimately I don't think JOSS can (or should) rule on package naming but I would like to consult with the broader JOSS editorial team on this point

@arfon sure, that sounds reasonable. Another thing to consider would be a clear statement in the readme of the project and the paper title that references the software created by @david-worrall .
