### 2021-05-24 #2446 @lockepatton
We are happy to add a statement at the top of our README disambiguating `SoniPy` and `soni-py` and point people to David Worrall's work, and have done so!

From what I'm gathering from @faroit and @afron, it sounds like sticking with `soni-py`, with a disambiguation in the README sounds like a good place to settle. Am I missing anything else?  If not, shall we unpause? Thanks!
