### 2021-05-24 #2446 @arfon
> Am I missing anything else? If not, shall we unpause? Thanks!

@lockepatton – are you planning on addressing this feedback from @faroit too? 👇

> I think it would be better to go with underscores, also otherwise this would not reflect a true renaming. See https://github.com/asteroid-team/asteroid-filterbanks as a good example of dash-hyphen in use at github/packaging/pypi
