### 2021-05-25 #2446 @lockepatton
Oh thanks for the reminder - I was scrolling through convinced I'd mentioned this before but it must have slipped my mind.

I'm very hesitant to republish anything on pypl simply because pypl doesn't remove old publications so the old `sonipy` will still exist. By renaming it, we'd be clogging up the `sonipy` namespace even more and risking people run `pip install sonipy` instead of `pip install soni_py`, thus installing the older version of two versions of the software or worse installing both. Since `SoniPy` and `soni-py` are distinct (both in dashes and capitalization), I'd like to leave it as is with `pip install sonipy` as our install command and titles including soni-py. Does this seem reasonable to you all?
