### 2021-05-25 #2446 @faroit
> by renaming it, we'd be clogging up the sonipy namespace even more and risking people run pip install sonipy instead of pip install soni_py, thus installing the older version of two versions of the software or worse installing both.

there are good practices for these issues: See [here for pypi](https://www.python.org/dev/peps/pep-0423/#how-to-rename-a-project). Also, this seems to make things a bit easier: https://github.com/simonw/pypi-rename

> I'd like to leave it as is with pip install sonipy as our install command and titles including soni-py. Does this seem reasonable to you all?

I would encourage you to try at least all possible ways to make the hyphenized name as effective as possible – given that these names are still quite close to the project of @david-worrall. Therefore, I would prefer the use of soni_py and soni-pi for the repo name, pypi package and the root src folder of the repo.

@lockepatton as @arfon was pointing out, we can't decide if renaming should take place and we will have to wait for other editors to jump in, so maybe just hold on and wait for more opinions on this matter?
