+++ paper/latexformat/paper.tex
@@ -70,9 +70,9 @@
 % \setlength{\parindent}{4em}
 \setlength{\parskip}{0.5em}
 
-\title{\texttt{sonipy}: A Perceptually Uniform Sonification Package}
+\title{\texttt{soni-py}: A Pitch-based Data Sonification Package}
 
-\author{Locke Patton \textsuperscript{123}, Emily Levesque \textsuperscript{14}}
+\author{Locke Patton \textsuperscript{1,2}, Emily Levesque \textsuperscript{1}}
 % \pdfauthor{Richard Southern}
 
 \keywords{sonification}
@@ -104,7 +104,15 @@
 
 \section{Introduction} \label{sec:introduction}
 
-\texttt{Sonipy} moves beyond visual analyses by sonifying scatter-plot data, producing audio files that depict variations in y as perceptually uniform changes in pitch. Short tones - called blips - are sounded in time at intervals corresponding to x values.
+\texttt{soni-py} moves beyond visual analyses by sonifying scatter-plot data, producing audio files that depict variations in y as perceptually uniform changes in pitch. Short tones - called blips - are sounded in time at intervals corresponding to x values.
+
+The addition of this audio sonification to a scientist's toolset has the intention of creating more inclusive and accessible science, new attention grabbing public outreach opportunities, and an easier self-consistent scientific understanding of minor differences in data over large scales.
+
+% The guarantees the user self consistency when translating from a visual medium to an audio medium. 
+
+% The addition of audio light curves might allow for more inclusive science, viral public outreach, as well as a better scientific understanding of the minor differences in objects. - benjamin rose
+
+% This specific case misses most of the examples in the JOSS documentation. However, I feel like this project could have a major impact on the field. In the study of odd astrophysical transients, a field that will see exponential growth in the next decade, light curve comparisons are the main scientific figure. This project has the potential to add audio “light curves” with fewer lines of code than you typically use to reformat a default matplotlib figure. The addition of audio light curves might allow for more inclusive science, viral public outreach, as well as a better scientific understanding of the minor differences in objects.
 
 \subsection{Understanding pitch}
 The cent is a logarithmic unit of measure for pitch intervals where $n \approx 3986\log(b/a)$ defines the number of cents between the pitch frequencies a and b.
@@ -114,16 +122,43 @@ The average person is capable of discerning independent subsequent pitches with 
 
 With these parameters, xy scatterplot data can be translated into audio files that map y values to specific pitch frequencies, with the minimum discernible $\Delta y$ corresponding to a 10 cent pitch difference.
 
-\section{The Case for Sonification}
+% \section{The Case for Sonification}
 
 %\subsection{Why sonify lightcurves?}
 
-Thanks to the nature of human hearing, we can audibly discern subsequent pitch differences of 10 cents. On a y scale ranging from 0 to 10, that corresponds to hearing variations as small as $\Delta y\sim0.03$. This simultaneous depth and range makes pitch-varied audio an incredibly powerful and accessible tool for understanding nuances in data. Through our sonification efforts of astronomical data we have found that listeners can discern differences in the brightness, duration, and period of time-domain variations when the changes in brightness are expressed as changes in pitch.
+% Thanks to the nature of human hearing, we can audibly discern subsequent pitch differences of 10 cents. On a y scale ranging from 0 to 10, that corresponds to hearing variations as small as $\Delta y\sim0.03$. This simultaneous depth and range makes pitch-varied audio an incredibly powerful and accessible tool for understanding nuances in data. Through our sonification efforts of astronomical data we have found that listeners can discern differences in the brightness, duration, and period of time-domain variations when the changes in brightness are expressed as changes in pitch.
 
[…]
