### 2021-05-26 #2446 @arfon
> I would encourage you to try at least all possible ways to make the hyphenized name as effective as possible – given that these names are still quite close to the project of @david-worrall. Therefore, I would prefer the use of soni_py and soni-pi for the repo name, pypi package and the root src folder of the repo.

I would also like to see this if possible. I went browsing the source code for your repository @lockepatton and as far as I could tell, it looked like the README had been modified but not a whole bunch more (caveat, I don't know very much about Python packaging).

> @lockepatton as @arfon was pointing out, we can't decide if renaming should take place and we will have to wait for other editors to jump in, so maybe just hold on and wait for more opinions on this matter?

Thanks for the suggestion @faroit but I don't think we should wait further. I've consulted with the broader JOSS editorial team and they agree that we ultimately can't mandate name changes by authors. That said, in the spirit of reducing user confusion as much as possible here, I would strongly encourage you @lockepatton to be as comprehensive as possible with this renaming i.e., following some of these suggestions 👇

> there are good practices for these issues: See here for pypi. Also, this seems to make things a bit easier: https://github.com/simonw/pypi-rename
