### 2021-06-21 #2446 @lockepatton
Hi @arfon and @faroit - it seems like the only remaining issue is resolving and making the name changes where needed.  To recap, I've changed the name to soni-py in the following places:
- README.md
- ontributing.md
- Tutorial.ipynb
- Paper
- tests README.md

I haven't changed the name in the GitHub repo to lockepatton/soni-py because I believe this'll hurt the efficacy of the code - those who have downloaded this code already will be met with a repository not found error upon attempts to connect with GitHub (updating the code for instance) - something I'd rather not cause unnecessarily to the entire user base. Even before publication, 16 people have already starred the code before publication - and I've advertised this and taught it in groups where this is their first interactions with GitHub and would like to minimize the difficulty of interacting with the code as much as possible.

I also haven't changed it on pip, as mentioned above, because the original "pip install sonipy" will not be removed. Two separate namespaces cause all sorts of issues including duplicate installations on the same machine, installing old versions unnecessarily and a rather clogged namespace on pip. This also asks all of our current users to uninstall sonipy and re-install soni-py unnecessarily.

In short, I'm happy to make any changes that are necessary to resolve the name issue but not want to do at the expense of the usability of the code.

With this in mind, are the current changes satisfactory? Do we need anything else before moving forward with the code?

(@arfon  The most update version of the code has major changes ending around 2 months ago, and minor changes up to today.)
