### 2021-06-22 #2446 @faroit
@lockepatton thanks for the update.

> I also haven't changed it on pip, as mentioned above, because the original "pip install sonipy" will not be removed. Two separate namespaces cause all sorts of issues including duplicate installations on the same machine, installing old versions unnecessarily and a rather clogged namespace on pip. This also asks all of our current users to uninstall sonipy and re-install soni-py unnecessarily.

Can you further elaborate on why you think that [pep-0432](https://www.python.org/dev/peps/pep-0423/#how-to-rename-a-project) recommendations wouldn't be a good solution here? I don't see that this would require users to reinstall the package...
