### 2021-07-06 #2446 @arfon
> Can you further elaborate on why you think that pep-0432 recommendations wouldn't be a good solution here? I don't see that this would require users to reinstall the package...

Yes, please address this point @lockepatton.
