### 2021-10-01 #2446 @arfon
@lockepatton – this review seems stuck. As @faroit suggested, the [pep-0432](https://www.python.org/dev/peps/pep-0423/#how-to-rename-a-project) seem like a good solution here. Could you please share your intent at this point?
