### 2021-11-03 #2446 @arfon
@lockepatton – I'm sorry to do this but if we don't hear a firm plan from your regarding this submission in the next two weeks, including addressing @faroit's suggestion around [pep-0432](https://www.python.org/dev/peps/pep-0423/#how-to-rename-a-project), we will be forced to reject this submission.
