### 2021-11-12 #2446 @lockepatton
Hi @afron and @danielskatz, Apologies for the delay - I will look at the pep-0432 guidelines for changing sonipy to soni-py on pip, hopefully make the change and get back to you within the next couple weeks. Thank you both.
