### 2022-02-18 #2446 @arfon
It's now three months since we last heard back from the author on this submission. At this point it seems like there are a few options we have here:

- @lockepatton commits to a firm timeline for making the (pep-0432) changes requested here.
- @lockepatton withdraws their paper.
- JOSS rejects this submission due to lack of activity on the part of the author.

@lockepatton - how would you like to proceed here? If we don't hear back from you in the next week we will go ahead and reject this submission.
